backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.39
pydantic==2.11.4
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-dotenv==1.0.1
email-validator==2.1.1
yfinance==0.2.37
pandas==2.2.3
requests==2.31.0
alembic==1.13.2
pytest==8.2.2
